Converting Double to BigDecimal to prevent underflow

+++ src/induction/hmm/HMM.scala
@@ -1,7 +1,7 @@
 package induction.hmm
 
 import induction.util.OneBasedArray
-import induction.util.Math
+import induction.math.Math
 import scala.collection.mutable.HashMap
 
 class HMM(val sentence:String, val λ:ModelParameters) {
@@ -19,26 +19,26 @@ class HMM(val sentence:String, val λ:ModelParameters) {
   val T: Int = O.length
     
   /** Defines a summation over hidden state indices. */
-  def Σ_hiddenStates(f:(Int => Double)) = Math.summation(λ.hiddenStateIndices, f)//λ.hiddenStateIndices.foldLeft(0.0) ((runningSum,i) => runningSum + f(i))
+  def Σ_hiddenStates(f:(Int => BigDecimal)) = Math.summation(λ.hiddenStateIndices, f)//λ.hiddenStateIndices.foldLeft(0.0) ((runningSum,i) => runningSum + f(i))
     
   /** Defines a summation over time, for t=1 to T-1. */
-  def Σ_time(f:(Int => Double)) = Math.summation(1 until T, f)
+  def Σ_time(f:(Int => BigDecimal)) = Math.summation(1 until T, f)
 
   /** Defines a summation over time, for t=1 to T. */
-  def Σ_observations(f:(Int => Double)) = Math.summation(1 to T, f)  
+  def Σ_observations(f:(Int => BigDecimal)) = Math.summation(1 to T, f)  
   
   /** Forward probability cache */
-  private val previouslyCalculated_α = new HashMap[(Int,Int),Double]
+  private val previouslyCalculated_α = new HashMap[(Int,Int),BigDecimal]
   
   /** Backward probability cache */
-  private val previouslyCalculated_β = new HashMap[(Int,Int),Double]
+  private val previouslyCalculated_β = new HashMap[(Int,Int),BigDecimal]
 
-  private val previouslyCalculated_ξ = new HashMap[(Int,Int,Int),Double]
+  private val previouslyCalculated_ξ = new HashMap[(Int,Int,Int),BigDecimal]
   
-  private val previouslyCalculated_γ = new HashMap[(Int,Int),Double]
+  private val previouslyCalculated_γ = new HashMap[(Int,Int),BigDecimal]
   
   /** Forward probability of entire observation sequence. */
-  def α : Double = {
+  def α : BigDecimal = {
     return Σ_hiddenStates( i => α(T,i) )
   }
 
@@ -48,7 +48,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * @param t Word index (1-based index)
    * @param j Hidden state value (1-based index)
    */
-  def α(t:Int,j:Int) : Double = {
+  def α(t:Int,j:Int) : BigDecimal = {
     
     if (previouslyCalculated_α.contains(t,j)) {
 
@@ -78,7 +78,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
   }
 
   /** Backward probability of entire observation sequence. */
-  def β : Double = Σ_hiddenStates( i => β(1,i) )
+  def β : BigDecimal = Σ_hiddenStates( i => β(1,i) )
   
   /**
    * Backward probability of the observation sequence after time <code>t</code>.
@@ -86,7 +86,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * @param t Word index (1-based index)
    * @param j Hidden state value (1-based index)
    */
-  def β(t:Int,j:Int) : Double = {
+  def β(t:Int,j:Int) : BigDecimal = {
      
     if (previouslyCalculated_β.contains(t,j)) {
 
@@ -102,7 +102,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
 		    } else if (j < 0 || j >= N) {
 		      throw new ArrayIndexOutOfBoundsException(j)
 		    } else if (t==T) {
-		      1.0
+		      BigDecimal(1)
 		    } else {
 		      Σ_hiddenStates( i => a(i,j) * b(j,O(t+1)) * β(t+1,j) )
 		    }
@@ -118,7 +118,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * Probability of being in state <code>i</code> at time <code>t</code>
    * and transitioning into state <code>j</code> at time <code>t+1</code> 
    */
-  def ξ(t:Int, i:Int,j:Int) : Double = {
+  def ξ(t:Int, i:Int,j:Int) : BigDecimal = {
     if (previouslyCalculated_ξ.contains(t,i,j)) {
       return previouslyCalculated_ξ(t,i,j)
     } else {
@@ -131,7 +131,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
   /**
    * Probability of being in state <code>i</code> at time <code>t</code> 
    */  
-  def γ(t:Int, i:Int) : Double = {
+  def γ(t:Int, i:Int) : BigDecimal = {
     if (previouslyCalculated_γ.contains(t,i)) {
       return previouslyCalculated_γ(t,i)
     } else {
@@ -152,7 +152,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * Returns the count of how often this HMM
    * is expected to begin in state <code>i</code>.
    */
-  def expectedStartsFrom_i(i:Int) : Double = {
+  def expectedStartsFrom_i(i:Int) : BigDecimal = {
     return γ(1,i)
   }
 
@@ -161,7 +161,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * is expected to transition from state <code>i</code>
    * into state <code>j</code>.
    */  
-  def expectedTransitionsFrom_i_to_j(i:Int, j:Int) : Double = {
+  def expectedTransitionsFrom_i_to_j(i:Int, j:Int) : BigDecimal = {
     return Σ_time( t => ξ(t,i,j) )
 //    var sum = 0.0
 //	for (t <- 1 to T) {
@@ -175,7 +175,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * is expected to transition from state <code>i</code>
    * into any next state.
    */   
-  def expectedTransitionsFrom_i(i:Int) : Double = {
+  def expectedTransitionsFrom_i(i:Int) : BigDecimal = {
       return Σ_time( t => γ(t,i) )
 //    var sum = 0.0
 //	for (t <- 1 to T) {
@@ -189,7 +189,7 @@ class HMM(val sentence:String, val λ:ModelParameters) {
    * is expected to emit observation <code>k</code>
    * from state <code>i</code>.
    */   
-  def expectedObservationsOf_k_from_i(i:Int, k:Int) : Double = {
+  def expectedObservationsOf_k_from_i(i:Int, k:Int) : BigDecimal = {
       return Σ_observations( t => {
           if (k == V.getInt(O(t))) 
               γ(t,i)
