+++ src/induction/mutable/ConditionalProbabilityDistribution.scala
@@ -4,20 +4,20 @@ import scala.collection.mutable.ArrayBuffer
 
 class ConditionalProbabilityDistribution(val variableVocabSize:Int, val givenVariableVocabSize:Int) extends induction.traits.ConditionalProbabilityDistribution {
 
-  val probabilities = new Array[Double](variableVocabSize * givenVariableVocabSize)
+  val probabilities = new Array[BigDecimal](variableVocabSize * givenVariableVocabSize)
   
   this.setUniform(0.0)
   
-  def update(givenAnotherVariable:Int, variable:Int, probability:Double) : Unit = {
+  def update(givenAnotherVariable:Int, variable:Int, probability:BigDecimal) : Unit = {
       val index = getIndexFor(variable,givenAnotherVariable)
     probabilities(index) = probability
   }
   
   def setUniform {
-    setUniform(1.0 / variableVocabSize.toDouble)
+    setUniform(BigDecimal(1) / BigDecimal(variableVocabSize))
   }
   
-  def setUniform(probability:Double) {
+  def setUniform(probability:BigDecimal) {
     for (i <- 0 until probabilities.size) { 
     	probabilities(i) = probability
   	}
@@ -34,7 +34,7 @@ object ConditionalProbabilityDistribution {
   def uniform(variableVocabSize:Int, givenVariableVocabSize:Int) : ConditionalProbabilityDistribution = {
     val d = new ConditionalProbabilityDistribution(variableVocabSize,givenVariableVocabSize)
     
-    d.setUniform(1.0 / variableVocabSize.toDouble)
+    d.setUniform(BigDecimal(1) / BigDecimal(variableVocabSize))
     
     return d
   }
